여러까지 ì예외처리 및 권한 체크 빡ì세게

+++ src/main/java/com/example/jsp_2023_06_am/ArticleDeleteServlet.java
@@ -18,6 +18,12 @@ public class ArticleDeleteServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         response.setContentType("text/html; charset=UTF-8");
+        HttpSession session = request.getSession();
+
+        if(session.getAttribute("loginedMemberId") == null){
+            response.getWriter().append("<script> alert('로그인 후 이용해주세요!!'); location.replace('../member/login');</script>");
+            return;
+        }
 
         Connection conn = null;
         try {
@@ -28,6 +34,16 @@ public class ArticleDeleteServlet extends HttpServlet {
             int id = Integer.parseInt(request.getParameter("id"));
 
             SecSql sql = new SecSql();
+            sql.append("select * from");
+            sql.append("article");
+            sql.append("WHERE id = ?", id);
+            Map<String, Object> articleMap = DBUtil.selectRow(conn, sql);
+
+            if (articleMap.isEmpty()){
+                response.setContentType("text/html; charset=UTF-8");
+                response.getWriter().append(String.format("<script> alert('해당 게시글이 존재하지 않습니다!!'); location.replace('list');</script>", id));
+                return;
+            }
             sql.append("DELETE FROM");
             sql.append("article");
             sql.append("WHERE id = ?", id);
